requirements.txt
transformers
evaluate
python-dotenv
torch
trl
peft
ipykernel
fire
datasets
deepspeed
bitsandbytes
accelerate
